requirements.txt
streamlit
streamlit-pydantic
pydantic==1.10.14
st-pages
python-dotenv
openai
llama-index
llama-index-core
llama-index-llms-openai
nltk
psycopg2-binary
requests
streamlit-folium
sqlalchemy
geocoder
geopy
pandas
BeautifulSoup4
googlesearch-python
spacy
https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.7.1/en_core_web_sm-3.7.1.tar.gz